refactor : BlackPawn과 WhitePawn 분리
