+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,7 +1,8 @@
 package chess.domain.piece;
 
+import chess.domain.move.BlackPawnMoveStrategy;
 import chess.domain.move.MoveStrategy;
-import chess.domain.move.PawnMoveStrategy;
+import chess.domain.move.WhitePawnMoveStrategy;
 
 public final class Pawn extends ValidPiece {
 
@@ -15,7 +16,10 @@ public final class Pawn extends ValidPiece {
 
     @Override
     public MoveStrategy getMoveStrategy() {
-        return new PawnMoveStrategy();
+        if (getColor() == Color.BLACK) {
+            return new BlackPawnMoveStrategy();
+        }
+        return new WhitePawnMoveStrategy();
     }
 
     @Override
